pyproject.toml
[tool.poetry]
name = "meals-api"
version = "0.1.0"
description = ""
authors = ["ticochuck <chucklivilla@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
django-restframework = "^0.0.1"
djangorestframework = "^3.12.1"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
